More details in guidelines
This includes updates from a recent change in txzchk and more details
about the tools and what they check (warnings and errors) and how to use
them individually.

This does need to be reviewed further, possibly being shorted or
referring to an FAQ, but it's a good start.

The rules had a typo fix as well.