Typo fix in SECURITY.md
Also removed what seems to be unnecessary quotes, much along the lines
in some FAQ links (that will be having the quotes removed).

Added SECURITY.md to bin/md2html.cfg and updated bin/gen-top-html.sh to
build SECURITY.html. The configuration can be improved for keywords and
other things and the file is not linked into the website either but
presumably it should be.