Fixes in bugs.{md,html}
The text 'If you want to try and fix this (mis)feature, you are welcome
to try' was problematic because these things are not (mis)features but
bugs. Thus it now reads '... fix this, ...'.

Also in one case that was noticed it was not actually anything that
should be fixed at all as they were documented features, not bugs.