Copy 1992/nathan.orig.c to nathan.alt.c
This is so one can see what the originally published code did. The
index.html has a bold important note that one should NOT send an email
to the address as it is not only no longer necessary but it is not known
if it is still valid. Unlike the index.html where the email was
omitted/redacted the alt code still has the email as the original code
as it is an exact copy thereof, including the typos in it.

Rebuilt .entry.json and index.html.

Updated .gitignore.