Fix and expand markdown guidelines
Improved wording of some of the titles.

Added div ids to each guideline so that we can link directly to them if
necessary.

The code block examples now specify a language so that when language
highlighting is available it will show up better. I added this for those
that are not wanted and those that are in the hopes it will stand out
more.

I added a guideline that for italics one should use underscores (_) as
using single asterisks is known to cause some parsers problems (and this
goes more so, perhaps, when it is bold italic text which would otherwise
be three asterisks in a row). I first became aware of this from a
document on some markdown parsers and it seems like a good idea to note.

Added that it is preferable and easier to use markdown links if possible
(along the lines of how it's easier and preferable to use inline code
blocks instead of spans). This would also help with consistency.

Typo fixes (some in code blocks, perhaps some outside as well).

For tabs/spaces discussion I added a tip for vim users. Maybe this
should be in the FAQ instead but I am not sure.

Move rule about not including literal parentheses in markdown links to
right below the rule about not ending markdown links with '))'.