Dead URL fixes in bugs file
That is make inline code blocks to help them stand out a bit more
(previously they were just normal text).

Another entry in this file just linked to another entry by the same
authors and not even to the discussion in the bugs file but just the
source code. This has been corrected too.