add bin/jval-wrapper.sh tool
We form a wrapper for the `jval(1)`, a tool that allow the printing of JSON values.

We propose that the grammar used by JSONPath.sh(1):

>   https://github.com/lcn2/JSONPath.sh

becomes the model for how `jval(1)` prints JSON values.  Consider, for, example:

>   JSONPath.sh -w -f 1984/mullender/.entry.json '$author_set..author_handle'

and in particular:

>   JSONPath.sh -w -b -f 1984/mullender/.entry.json '$author_set..author_handle'

as a model of printing the `author_handle` values from the JSON `author_set` array.

See the following URL for more information.

>   https://github.com/lcn2/JSONPath.sh?tab=readme-ov-file#jsonpath-patterns-and-extensions