speed up make quick_www
We add `-Q` to `bin/gen-other-html.sh` to add a mode
that will only update an entry non-index.html if it
is missing or out of date.

Added `make quick_other_html` that calls `bin/gen-other-html.sh`
with the `-Q` quick mode option.

Changed `bin/gen-other-html.sh` VERSION from "1.2.6 2024-08-05"
to "1.3 2024-08-11".