add temporary summary.csv file
This file will go away, but is a good start for the
title and abstract for entries.

The contents started from
[summary.txt](https://github.com/ioccc-src/winner/blob/master/all/summary.txt).

We converted the contents into a CSV format.  Moreover the abstract was
limited to no more than 63 characters.  All comma's (",") within the
abstract were removed to make converting to CSV easier.

NOTE: This is a tempory file that will **GO AWAY** once the all of the
`.entry.json` files have been updated with a title and abstract.