Try (yet again) fixing workflow test.yml
It seems that the problem was (or perhaps not .. we'll see :-) ) is that
the 'run:' part of each task in the workflow file should not have a '-'.
If this is the case then this should solve the problem and if so the
problem was that some had it and others had it and in a rush I thought,
forgetting the syntax, that the problem was the missing ones were
incorrect.