Reorder FAQ table of contents
Numerous sections have had questions moved to another section including
the new sections 'The mkiocccentry toolkit: finer details' and 'History
of the IOCCC'.

The first section was renamed (pending any better name as it might not
be quite right) 'Entering the IOCCC: the bare minimum you need to know'.

Renamed some questions to better (hopefully) account for what they are
about.

The purpose of not making the actual content match this reordering is so
that if any changes need to be made they do not have to be done again
(though a few were moved). Thus before more work can be done some
discussion is needed.

The guidelines and rules now refer at the top to the general FAQ
especially about the bare minimum one needs to know.