Reorder FAQ table of contents, adding new sections
New sections were added. Not all anchors work yet. The questions/answers
are not yet in the right order. That is the next step.

Instead of relying on having to manually add numbers use ordered lists.
The sections are in unordered lists so we have control over the number.
The downside with this is with the long file it is harder to find the
section one is in but as each question/answer has a jump to the top that
should be worth the gain of not having to manually edit every question
(in the table of contents as well as the question/answers) every time a
question is moved to another place (or a new one is added) above others.

Another downside is minor: it does not allow for 0 as html lists do not
start with 0 but 1.

Another thing that has to be done is to make sure that the html lint is
correct. That can be done once this is merged. Any thing that needs to
be fixed can then be fixed.