fix minor FAQ issues, shortened some FAQ titles, added Q x.y
We fixed some of minor inconsistencies in the `faq.md` markdown.

We added "Q x.y:" to the Table of Contents and FAQ titles,
in part to verify the order, and in part to help someone find
their way while inside the FAQ content.

Reduced several FAQ titles that were wrapping onto a 2nd line
at the standard window width.

Perform `make faq` to perform the above and test under macOS.