Add missing file in 2001/rosten
The new file, greasymouse.man, IS the cat man page noted by the author
in their remarks. This file was provided by Leo as remark.txt but to
make it render better I have made it a .man file and, for syntax
highlighting, I moved the section titles to the first column of the
file.

I thought of making it a true man page but as the author suggested that
it is a cat man page I decided against this.