Important fixes in 2020/ferguson2
A typo in markdown meant that a link was not formed.

Renamed enigma.html to enigma.txt. Although it was to be set to viewed
as text the problem is that as a html file running a linter on it will
cause errors (obviously).

The .entry.json and all files that referred to the file enigma.html (not
in remote web servers, obviously, but just this website) now refer to
enigma.txt, except that enigma.html was now added to .gitignore and the
recode.md (in the example command) redirects to enigma.html as that
should be a html file, assuming they have the right key (of the
selection I gave).

These problems were not discovered due to trying to get the html files
in a good state (ironically) and knowing that I had written a lot.

I believe now that the files should be okay.

Although I did use sgit I did it carefully and I still ran make
quick-readme2index and make gen_other_html.