Fix html lint (W3C Nu) errors
In the case of two which had an abstract (.entry.json file) that
includes one (in one case) '<' or two (in another case) '<'s a markdown
code block had to be used as otherwise the abstract would have to have a
&lt; which looks even worse. This is not ideal but if we want it to pass
lint this is what has to be done. Use of json \uxxxx won't work as that
encodes it to < literally.

In the case of 2018/burton[12] the author had a section called Coda.
This conflicted with the problem that all generated html files have such
a section. To fix this problem I renamed it 'Final remarks'. Not ideal
but at least it now passes as valid.

There were a number of different kinds of problems and in one case the
symbols < and > could be fixed by changing it to a code block (which it
should have been anyway) - that was 2020/ferguson1 I think.