Remove `s from .entry.json files
The way the abstract is used for the index.html files does not allow for
parsing by pandoc so it did not work as expected (I did not verify the
two entries this is a problem).

The index.html files of these two entries have been modified. Unless the
script is updated the only other way is to update the abstract to use
&lt; but this can be discussed.