Add html.sed to translate certain characters
The script, for now, translates the following characters to their html
entity:

    &, <, >

An important note that is in a comment is that the & MUST come first as
otherwise the & in the others will be corrupted.

One script uses this but in order to make sure it is okay make www will
have to be run first. Another script, the one that needs it even more,
has a problem that has to be discussed, before it can be used.
Nonetheless the html.sed file is at least added for the scripts once
things are tested okay.