Add try.sh/try.alt.sh template scripts
The scripts can be found under template/entry. A new FAQ entry has been
added and the guidelines have been updated.

The rules have a new div id ("version") to match what was added to
guidelines yesterday (since it's obvious what kind of version it is.

Once this has been merged html files need to be checked to make sure
everything is okay.