Fix version in FAQ
The question of why did it jump from

    28.1.3 2024-10-25

to

    28.1.6 2024-12-05

can be explained as such. There were three edits since the last version
update so there should be three version updates. As for why 5 December
2024 when today is 6 December 2024: it's because the last update to the
contents was yesterday, 5 December 2024, although I believe there could
be some improvements (menu wise) for Makefile.example. This has to be
looked into and if it is done the version will be bumped again.