Add initial submit instructions to FAQ
It might be worth pointing out that the screenshots might be better
done. For instance some of the longer lines are in some cut off. This
was to try and fit it in smaller screens but it might not be desirable.
Another issue is the scaling of one of the smaller images might make it
too small, at least for people like me who are practically blind.

The wording could definitely be improved as well but I wanted this
initial update in so it can be looked at better and any improvements
necessary can be discussed.