Improve upload submission FAQ
New screenshots have been added with new explanations. A couple files
were either deleted or renamed.

There were some fixes as well.

I have looked at the result on the web server and it appears to look and
read well and hopefully this will suffice for this part of the how to
submit FAQ.