Update/improve two entries' abstracts
1985/sicherman now has an abstract of: rot13 w/strange #defines that
create comment headers etc.

I note that this is longer than what I believe was what was desired.
However it is the same length as it was before, only that it now has #defines
 instead of #defs.

1988/phillipps has: 12 Days of Christmas recursive main() calls w/tables
instead of 'days of christmas heavily main() calling with tables'. The
rewording was necessary to keep the length the same but to be more
clear. I just remembered however, that to fix it for modern systems, a
new function had to be added which is the recursive function, so it no
longer is strictly true that it it is recursive main() calls. It might
be better to keep it main(), however, as that was the original award.
The difference here is that it is now a bit more correct in the output.