Make IOCCC status text consistent with JSON status
The status showed 'The IOCCC is not yet open' to indicate it is in
pending status. But it would be better to match the actual status
itself especially as the status.html file explains pending status, not
'not yet open'. At a quick glance it is easy to miss especially for
those of us who have worked on the repo for a long time and in
particular those who worked on the status.html file a lot.

This updated files in inc/, bin/ and next/ but I cheated and used sgit
to do this. I then rebuilt the html files. A previous minor grammar
problem in rules was fixed as well.