Typo fix in rules and guidelines
This includes the include files inc/guidelines.pending.hdr and
inc/rules.pending.hdr.

Rebuilt html files as well.

This does not include the other pending fixes and it does not include
the updates about mkiocccentry now allowing subdirectories. That'll come
in another commit either later today or if not then hopefully this
weekend.