modify PR #54
Hello @tanaydin,

We are about to modify the result of your PR #54, @tanaydin to include
a link to the finger daemon source.  Unfortunately the finger daemon
source is not found in that 4.2BSD repo, so we had to dip into the
[4.3BSD-Reno](https://github.com/dank101/4.3BSD-Reno) repo, which isn't
exactly correct as that post-dates the 1984 IOCCC starting event but
will have to do.

We modified the link to use HTML "Left Parenthesis" and "Right Parenthesi"
as per [IOCCC markdown guidelines regarding LITERAL “(” or “)” in
markdown link titles](https://www.ioccc.org/markdown.html#parentheses)
in order to your mod to pass the web site test code.

We also modified the references in the FAQ, as per your suggestion.

We also performed needed changes to update the associated HTML files.

We thank you for recommending this change.

Stay tuned for the update.