Update FAQ, guidelines and rules
The new syntax (though not all functionality is yet finished) of
mkiocccentry does not match what was in the FAQ, guidelines and rules.

Also added an important reminder to use v7 format with tarballs as not
doing so can reveal who you are which would break the rules. This is
mentioned in the rules and guidelines, showing how to create such a
tarball.

More work does have to be done with these two things however.