Add important note about mkiocccentry execution
It is highly recommended that you run the tool from either the repo
directory itself (making sure you have the most up to date version of
course) or make sure you have installed it (also making sure you have
the most recent version). Not doing this will cause issues with checks
and in the case of out of date versions the submission would be rejected
due to violating rule 17. One can specify the options to mkiocccentry
for the tools required but if you forget or don't have the most up to
date version you will very likely run into problems.