Update FAQ for mkiocccentry process
I have thought of and plugged additional holes in the process (same with
txzchk but that's beyond the scope of the document) so the FAQ was
updated for this.

Also because it is a long list of steps there is now a TL;DR part above
it.

There was an order issue in the list as well.