Update pw-change.html and register.html
Some images were deleted as they are no longer needed (they are just
text - although based on discussion some of the text might change).

Adjusted alt text of the images that remain in these files.

The overall look does have to be checked more carefully along with the
rules. The guidelines still have to be checked.