Update register/submit/pw-change for screenreaders
The three important files, next/{register,submit,pw-change}.{md,html}
no longer have images in them (besides the logo at the top of the image
which has nothing to do with registering, submitting or changing
passwords); instead the text in the images are actual text (in some
cases, just the relevant parts and in others all or almost all).

The png/register.* and png/submit.* images were removed.